Fix inconsistency with phone_contacts output
Previously, I used model aliases to trick the front-end into displaying the value I wanted, but this caused JSON responses to have `phone` and HTML responses to have `phones`, which was confusing.

Now, the fields are always plural, and I use `inflection` to singularize the value for displaying on the summary page.