[EDS-556] Supports SAML sign-on to unlock other search options
When a user visits the page, they are now shown the sign in link in order to see other search options.

Fixes the mock idp integration so that it works locally and is testable.

Adds tests.

Note: Does **NOT** enable student search yet. The radio buttons won't currently have any effect on the search results.