Simplify the browser<->backend communication
The original intent of the backend API design was because we originally intended this to be a javascript-powered experience. However, once we realized it wasn't, we never addressed the back-end API, which resulted in a lot of extra complexity.

This change makes the front-end request input model match the actual front-end use case, which removes several additional validations, translations, and an extra javascript function.

The request input is now preserved exactly as-is and returned exactly as-was.

Additionally I have added tests to ensure that this behavior stays in sync.

The input model is now simply:

```
{
    "method": "name",  // or phone, boxNumber, etc.
    "query": "ada lovelace",  // actual query term
    "population": "all",  // or "students", "employees"
    "length": "summary" // or "full"
}
```

This correlates 1:1 with the form, so we no longer have to coerce these two models.

Additionally:

- Removes unused check for "show_count"
- Fixes the "has_tag_with_text" to fix false positives (replacement is heavily documented)