Add support for href field in PWS and Search models
PWS returns a convenient URL to get person results, which we can use to fetch data for vcard (as well as "More" queries).

The PWS model now supports setting the href from the PWS results, as well as the user regid, in case we need to construct our own href (PWS does not return the href when full results are queried)

The Search model now supports setting an href as well, but base64-encodes the value first, so that the front-end can can use the value without worrying about funny characters that PWS may add into their url(s) over time.