Update `Person` model to support departments and titles
Previously, the model only allowed for a single `department: str`, which is insufficient for us to reach functional parity with the C-based directory service. This updates the model so that the final result includes all "titles" and "departments" for various roles a person might have within the university.

Other changes:

- Optimizes imports