[EDS-563] Support department queries (#51)
* Adds support for department queries

Now allows users to search by department. Like the original directory, this will substitute "and" for "&" (and vice versa) to perform all relevant searches.

* Increases global test coverage to 99% and cleans up some crufty code

does what it says, no functional changes.

* [EDS-563] Export department and title information in JSON data. (#52)

* [minor] Refactors a common pattern in the tests

- Similar to the `mock_injected` fixture in `identity-uw` (if not exactly the same), this is a convenience method for creating global mocks of request-based instances of injected objects. This allows us to set things like session state that would otherwise require a full request context to be testable.

Additional changes:

- Optimizes imports
- Minor formatting updates
- Adds departments and titles to employee and student mock people.
- Adds a student_employee mock person.

* Update `Person` model to support departments and titles

Previously, the model only allowed for a single `department: str`, which is insufficient for us to reach functional parity with the C-based directory service. This updates the model so that the final result includes all "titles" and "departments" for various roles a person might have within the university.

Other changes:

- Optimizes imports

* Updates the PWS translator to fill department and title values

* Updates search service tests to assert department output.

Other changes:

- Refactors to be in line with other tests' use of `mock_injected`

* Ignore incomplete ("holey") position data from PWS