[EDS-568] Actually query for student data when selected (and auth'd) (#54)
Since the PWS search is intersectional, we can't search both for students and employees in a single query (that only returns 'student employees'). This change adds logic to the QueryGenerator to emit student population queries if the request input dictates students should be included, and if the request is authenticated.