[EDS-584][ICEBERG] Fix bug in search result filtering (#70)
**BIG CHANGE**

Closes EDS-584

This is a bigger change than you'd expect, because the fundamental architecture of the post-request filtering was awkwardly designed (see? It's not just others' code i refactor mercilessly.)

This fixes the problem by formalizing the post-request filtering into a single standard set of constraints, while also allowing certain queries to apply additional constraints.

Then `QueryGenerator`  was updated to add constraints around names, where if a preferred name is set, then the registered name must match.

Because of the iceberg nature of this change, you should pay particular attention to the constraints added to `services/query_generator.py`. If you want to understand constraints further, check out `husky_directory/models/common.py#RecordConstraint` — this will allow us to easily add and maintain filters based on queries, as well as application-wide filters, on the data we receive from PWS.

Another place that is useful to look is in `services/pws.py`, _particularly_ the `global_constraints` property, which defines application-wide constraints we will place on all data.

Now all of our constraints are handled the same way by a central component, as close to the "raw" data as possible.

The actual test to prove EDS-584 can close is the test in `tests/blueprints/test_search_blueprint.py`: `test_get_person_non_matching_surname`.

This also re-scopes several modules to the `request` scope, instead of the `singleton` scope, to further add to application safety.