Use a simpler approach to create release images.
Lately a lot of work has gone into the deploy process, but the release process represented in automation
hasn't really caught up. As a result, some old ideals were being used which were causing
release images to be built without the properly set variables.

This update simplifies the workflow by separating the "github release" from the "docker push".
Doing so, we no longer have to try and resolve a pull-request as part of this workflow,
and instead check out the release tag when creating the docker image.

This saves us from accidentally introducing drift in what is released vs. what is running,
and unifies the process using the already ubiquitous `pre-push.sh`.