Use a simpler approach to create and push release images. (#113)
* Add some control toggles to `pre-push.sh`

- `--no-commit` avoids trying to make any commits if auto-formatting code (automatically set when run in `--test` mode)
- `--skip-auto-format/-k` avoids automatically reformatting code
     but does _not_ skip the checks that validate code format.
     (Useful when `black` isn't installed and no commits should be made.)

* Don't skip release job if running in dry-run, even if it wouldn't normally be required

* Use a simpler approach to create release images.

Lately a lot of work has gone into the deploy process, but the release process represented in automation
hasn't really caught up. As a result, some old ideals were being used which were causing
release images to be built without the properly set variables.

This update simplifies the workflow by separating the "github release" from the "docker push".
Doing so, we no longer have to try and resolve a pull-request as part of this workflow,
and instead check out the release tag when creating the docker image.

This saves us from accidentally introducing drift in what is released vs. what is running,
and unifies the process using the already ubiquitous `pre-push.sh`.