Another small fix/tweak for the release workflow
A race condition previously allowed the 'deploy' step to run before the image was actually pushed to GCR. Remedied that by moving the deploy step into the create-release workflow.

Also, added gcr login step to the job that pushes to docker, which i forgot before but couldn't test because of dry runs.